towards a nice split
