+++ domaindriven-server/src/DomainDriven/Server.hs
@@ -5,8 +5,10 @@
 module DomainDriven.Server
     ( mkServer
     , module DomainDriven.Server.Config
+    , module P
     )
 where
 
 import DomainDriven.Server.Config
+import qualified DomainDriven.Server.Param as P
 import DomainDriven.Server.TH
