+++ domaindriven-server/src/DomainDriven/Server/Config.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TypeFamilyDependencies #-}
 
 module DomainDriven.Server.Config
     ( module DomainDriven.Server.Config
@@ -8,11 +9,18 @@ module DomainDriven.Server.Config
 where
 
 import Data.Char (isLower)
+import Data.Function (on)
+import Data.Generics.Product
 import qualified Data.List as L
 import qualified Data.Map as M
+import Data.Proxy
 import DomainDriven.Internal.Class
+import DomainDriven.Server.Param
 import GHC.Generics (Generic)
+import GHC.TypeLits
 import Language.Haskell.TH
+import Lens.Micro ((^.))
+import UnliftIO (MonadUnliftIO)
 import Prelude
 
 -- | Configuration used to generate server
@@ -24,53 +32,34 @@ data ServerConfig = ServerConfig
     }
     deriving (Show, Generic)
 
-------------------------------------------------------------------------------------------
--- Things to deal with the P type family
-------------------------------------------------------------------------------------------
-type PAction = ParamPart -> Type -> Type -> Type
-
--- | Used as a parameter to the `P` type family on order to determine the focus.
-data ParamPart
-    = ParamName
-    | ParamType
-    deriving (Show)
-
--- | P is used for specifying the parameters of the model.
--- The name will be used as the name in the JSON encoding or the query parameter of the
--- generated server.
-type family P (b :: ParamPart) (name :: Symbol) (a :: Type) where
-    P 'ParamName name ty = Proxy name
-    P 'ParamType name ty = ty
-
-runPAction
-    :: (MonadUnliftIO m, WriteModel p, model ~ Model p, event ~ Event p)
-    => p
-    -> ActionHandler model event m cmd
-    -> cmd 'ParamType method ret
-    -> m ret
-runPAction p handleCmd cmd = case handleCmd cmd of
-    Query m -> m =<< liftIO (getModel p)
-    CbQuery m -> m (liftIO (getModel p))
-    Cmd m -> transactionalUpdate p m
-    CbCmd withTrans -> withTrans $ \runTrans -> do
-        transactionalUpdate p runTrans
-
-type PActionHandler model event m c =
-    forall method a. c 'ParamType method a -> HandlerType method model event m a
-
-type PActionRunner m c =
-    forall method a
-     . MonadUnliftIO m
-    => c 'ParamType method a
-    -> m a
-
-------------------------------------------------------------------------------------------
--- Done with the P type family things
-------------------------------------------------------------------------------------------
---
 class HasApiOptions (action :: ParamPart -> Type -> Type -> Type) where
     apiOptions :: ApiOptions
     apiOptions = defaultApiOptions
+
+data ApiOptions = ApiOptions
+    { renameConstructor :: String -> String
+    , typenameSeparator :: String
+    , bodyNameBase :: Maybe String
+    }
+    deriving (Generic)
+
+defaultApiOptions :: ApiOptions
+defaultApiOptions =
+    ApiOptions
+        { renameConstructor =
+            \s -> case L.splitAt (on (-) L.length s "Action") s of
+                (x, "Action") -> x
+                _ -> s
+        , typenameSeparator = "_"
+        , bodyNameBase = Nothing
+        }
+
+instance Show ApiOptions where
+    show o =
+        "ApiOptions {renameConstructor = ***, typenameSeparator = \""
+            <> o
+                ^. field @"typenameSeparator"
+            <> "\"}"
 
 defaultServerConfig :: ServerConfig
 defaultServerConfig = ServerConfig M.empty
